@@ -14,13 +14,14 @@
     "materialize-css": "^1.0.0-rc.2",
     "moment": "^2.29.0",
     "path-to-regexp": "^1.7.0",
-    "react": "^16.13.1",
+    "react": "^16.14.0",
     "react-chartjs-2": "^2.10.0",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.1",
+    "react-dom": "^16.14.0",
+    "react-redux": "^7.2.6",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
+    "react-table": "^7.7.0",
     "redux": "^4.0.5"
   },
   "scripts": {
