@@ -33,11 +33,10 @@
     "jsonwebtoken": "^8.5.1",
     "materialize-css": "^1.0.0-rc.2",
     "mongoose": "^5.10.6",
-    "nouislider": "^14.6.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-nouislider": "^2.0.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-redux": "^7.2.1",
+    "react-window": "^1.8.6",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "wnumb": "^1.2.0"
