@@ -1,7 +1,7 @@
 import playerListReducer from './playerListReducer'
 import authReducer from './authReducer.js'
 import squadListReducer from './squadListReducer.js'
-import bidReducer from './bidReducer.js'
+import shortListReducer from './shortListReducer.js'
 
 import {combineReducers} from 'redux'
 
@@ -9,7 +9,7 @@ const allReducers = combineReducers({
 	playersList: playerListReducer,
 	auth: authReducer,
 	squadList: squadListReducer,
-	bidState: bidReducer
+	shortState: shortListReducer
 })
 
 export default allReducers
