@@ -6,70 +6,69 @@ const schema = new Schema({
 	nation: {type: String},
 	ca: {type: Number},
 	pa: {type: Number},
+	preferredfoot: {type: String},
 	age: {type: Number},
 	position: {type: String},
 	height: {type: String},
 	weight: {type: String},
-	cr: {type: Number},
-	preferredfoot: {type: String},
-	mins: {type: Number},
-	avrat: {type: Number},
+	wpneeded: {type: String},
 	price: {type: Number},
+	status: {type: String},
 	club: {type: String},
-	lth: {type: String},
-	lon: {type: String},
-	dri: {type: String},
-	fin: {type: String},
-	hea: {type: String},
-	fre: {type: String},
-	cro: {type: String},
-	mar: {type: String},
-	tck: {type: String},
+	lth: {type: Number},
+	lon: {type: Number},
+	dri: {type: Number},
+	fin: {type: Number},
+	hea: {type: Number},
+	fre: {type: Number},
+	cro: {type: Number},
+	mar: {type: Number},
+	tck: {type: Number},
 	pas: {type: Number},
-	pen: {type: String},
-	fir: {type: String},
-	tec: {type: String},
-	cor: {type: String},
-	agg: {type: String},
-	vis: {type: String},
-	pos: {type: String},
-	otb: {type: String},
-	fla: {type: String},
-	ant: {type: String},
-	tea: {type: String},
-	cnt: {type: String},
-	ldr: {type: String},
-	dec: {type: String},
-	wor: {type: String},
-	det: {type: String},
-	cmp: {type: String},
-	bra: {type: String},
-	sta: {type: String},
-	jum: {type: String},
-	bal: {type: String},
-	agi: {type: String},
-	nat: {type: String},
-	str: {type: String},
-	pac: {type: String},
-	acc: {type: String},
-	amb: {type: String},
-	impm: {type: String},
-	prof: {type: String},
-	injpr: {type: String},
-	cons: {type: String},
-	thr: {type: String},
-	com: {type: String},
-	kic: {type: String},
-	cmd: {type: String},
-	tro: {type: String},
-	han: {type: String},
-	ovo: {type: String},
-	pun: {type: String},
-	pasgk: {type: String},
-	firgk: {type: String},
-	ref: {type: String},
-	aer: {type: String},
-	ecc: {type: String}
+	pen: {type: Number},
+	fir: {type: Number},
+	tec: {type: Number},
+	cor: {type: Number},
+	agg: {type: Number},
+	vis: {type: Number},
+	pos: {type: Number},
+	otb: {type: Number},
+	fla: {type: Number},
+	ant: {type: Number},
+	tea: {type: Number},
+	cnt: {type: Number},
+	ldr: {type: Number},
+	dec: {type: Number},
+	wor: {type: Number},
+	det: {type: Number},
+	cmp: {type: Number},
+	bra: {type: Number},
+	sta: {type: Number},
+	jum: {type: Number},
+	bal: {type: Number},
+	agi: {type: Number},
+	nat: {type: Number},
+	str: {type: Number},
+	pac: {type: Number},
+	acc: {type: Number},
+	amb: {type: Number},
+	impm: {type: Number},
+	prof: {type: Number},
+	injpr: {type: Number},
+	cons: {type: Number},
+	thr: {type: Number},
+	com: {type: Number},
+	kic: {type: Number},
+	cmd: {type: Number},
+	tro: {type: Number},
+	han: {type: Number},
+	ovo: {type: Number},
+	pun: {type: Number},
+	pasgk: {type: Number},
+	firgk: {type: Number},
+	ref: {type: Number},
+	aer: {type: Number},
+	ecc: {type: Number}
 })
 
 module.exports = model('Shortlist', schema)
