@@ -35,10 +35,11 @@ router.get('/loadsquad', (req, res) => {
 				});
 })
 
-/* ------------------------------------ */
-/* Возвращает список всех бидов в базе */
-/* ---------------------------------- */
+/* ------------------------------------------------------ */
+/* Возвращает список игроков клуба добавленных в шортлист */
+/* ------------------------------------------------------ */
 router.get('/loadbid', (req, res) => {
+		/* query { club: clubName } */
 		Shortlist.find(req.query)
 				.then((data) => {
 						res.json(data);
@@ -48,76 +49,10 @@ router.get('/loadbid', (req, res) => {
 				});
 })
 
-/* ---------------------------------- */
-/* Возвращает, текущий бид на игрока */
-/* -------------------------------- */
-router.get('/loadcurrentbid', (req, res) => {
-		Bid.findOne(req.query)
-				.then((data) => {
-						res.json(data);
-				})
-				.catch((error) => {
-						console.log('error: ', error);
-				});
-})
-
-/* ---------------------------------------- */
-/* Добавляет или изменяет существующий бид */
-/* -------------------------------------- */
-router.post('/bidsend', (req, res) => {
-	const data = req.body
-	const filter = req.query
-	const options = {upsert: true, new: true, useFindAndModify: false, rawResult: true}
-
-	Bid.findOne(filter, 
-		(error, doc) => {
-		if (error) {
-				res.status(500).json({ msg: 'Sorry, internal server errors' });
-				return 
-		}
-		if (!doc) { 
-			Bid.findOneAndUpdate(
-				filter, data, options, 
-				(error, doc) => {
-						if (error) {
-								res.status(500).json({ msg: 'Sorry, internal server errors' });
-								return; 
-						}
-						return res.json(doc);
-			})
-		}
-		else if (Number(data.curPrice) < Number(doc.nextPrice)) /* если присланный бид меньше, чем ожидаемый следующий бид */
-			return res.json('toLateError')
-		else { 
-
-			/* возвращаем деньги */
-			User.findOneAndUpdate({userId: doc.userId},  {$inc: {money: Number(doc.curPrice)}}, options,
-			(error) => {
-					if (error) {
-							res.status(500).json({ msg: 'Sorry, internal server errors' });
-							return; 
-					}
-			})  
-
-			data.prevBid = {userId: doc.userId, club: doc.club, price: doc.curPrice}
-			Bid.findOneAndUpdate(
-			filter, data, options, 
-			(error, doc) => {
-					if (error) {
-							res.status(500).json({ msg: 'Sorry, internal server errors' });
-							return; 
-					}
-					console.log('это был не первый бид на игрока, но я первее перебил');
-					return res.json(doc);
-			})}
-
-});
-});
-
 /* ----------------------------------- */
 /*     Добавление игрока в шортлист    */
 /* ----------------------------------- */
-router.post('/shortlistupdate', async (req, res) => {
+router.post('/shortlistadd', async (req, res) => {
 	const data = req.body
 	const newShortPlayer = await Shortlist.create(data)
 	res.json(newShortPlayer)
@@ -141,50 +76,52 @@ router.post('/shortlistremove', (req, res) => {
 /*     Отчисление игрока    */
 /* ----------------------- */
 router.get('/sellsquadplayer', (req, res) => {
-		const uid = req.query.uid
-		Squadlist.find(
-				{uid: {$nin: uid}, club: req.query.club},
-				(error, doc) => {
-						if (error) {
-								res.status(500).json({ msg: 'Sorry, internal server errors' });
-								return; 
-						}
-						return res.json(doc);
-				});
+	const uid = req.query.uid
+	Squadlist.find(
+		{uid: {$nin: uid}, club: req.query.club},
+			(error, doc) => {
+				if (error) {
+					res.status(500).json({ msg: 'Sorry, internal server errors' });
+					return; 
+				}
+				return res.json(doc);
+			}
+	);
 });            
 
 /* ------------------------------------- */
 /*     Обновляет деньги пользователя    */
 /* ----------------------------------- */
 router.post('/changeuser', (req, res) => {
-		const data = req.body
-		const filter = req.query
-		const options = {new: true, useFindAndModify: false}
+	const data = req.body
+	const filter = req.query
+	const options = {new: true, useFindAndModify: false}
 
-		User.findOneAndUpdate(filter,  {$inc: {money: Number(data.money)}}, options,
-				(error, doc) => {
-						if (error) {
-								res.status(500).json({ msg: 'Sorry, internal server errors' });
-								return; 
-						}
-						return res.json(doc.money); /* Возвращает только количество денег */
-				})  			
+	User.findOneAndUpdate(filter,  {$inc: {money: Number(data.money)}}, options,
+		(error, doc) => {
+			if (error) {
+				res.status(500).json({ msg: 'Sorry, internal server errors' });
+				return; 
+			}
+			return res.json(doc.money); /* Возвращает только количество денег */
+		})  			
 })
 
 router.post('/setbidstatus', (req, res) => {
-		const data = req.body
-		const filter = req.query
-		const options = {new: true, useFindAndModify: false}
+	const data = req.body
+	const filter = req.query
+	const options = {new: true, useFindAndModify: false}
 
 	Player.findOneAndUpdate(
-			filter, {$set: data}, options, 
-			(error, doc) => {
-					 if (error) {
-							 res.status(500).json({ msg: 'Sorry, internal server errors' });
-							 return; 
-					 }
-					 return res.json(doc);
-	 });
+		filter, {$set: data}, options, 
+		(error, doc) => {
+			if (error) {
+				res.status(500).json({ msg: 'Sorry, internal server errors' });
+				return; 
+			}
+			return res.json(doc);
+		}
+	);
 });
 
 // If no API routes are hit, send the React app
